Updates README with social links and banner
Enhances the project's README file by adding a banner image and a table of social media and related links.

This provides users with a more visually appealing and informative landing page for the project, and facilitates easier access to the project's online presence and resources.

Removes unused assets.
Adds funding model file.